update area, bar buble and linechart
