+++ src/ts/components/charts/BubbleChart.tsx
@@ -1,15 +1,16 @@
-import { BubbleChart as MantineBubbleChart } from "@mantine/charts";
+
 import { BubbleChartDataKey } from "@mantine/charts";
 import { MantineColor } from "@mantine/core";
 import { BoxProps } from "props/box";
 import { GridChartBaseProps } from "props/charts";
 import { DashBaseProps } from "props/dash";
 import { StylesApiProps } from "props/styles";
-import React from "react";
-import { getScatterClickData, isEventValid } from "../../utils/charts";
-import { getLoadingState } from "../../utils/dash3";
+import React, { Suspense } from "react";
 
-interface Props
+// eslint-disable-next-line no-inline-comments
+const LazyBubbleChart = React.lazy(() => import(/* webpackChunkName: "BubbleChart" */ './fragments/BubbleChart'));
+
+export interface Props
     extends BoxProps,
         StylesApiProps,
         DashBaseProps {
@@ -48,31 +49,12 @@ interface Props
 }
 
 
-/** ScatterChart */
+/** BubbleChart */
 const BubbleChart = (props: Props) => {
-    const { setProps, loading_state, clickData, hoverData, scatterProps, ...others } =
-        props;
-
-    const onClick = (ev) => {
-        if (isEventValid(ev)) {
-            setProps({ clickData: getScatterClickData(ev) });
-        }
-    };
-
-    const onMouseOver= (ev) => {
-        if (isEventValid(ev)) {
-            setProps({ hoverData: getScatterClickData(ev) });
-        }
-    };
-
-    const newProps = { ...scatterProps, onClick, onMouseOver};
-
     return (
-        <MantineBubbleChart
-            data-dash-is-loading={getLoadingState(loading_state) || undefined}
-            scatterProps={newProps}
-            {...others}
-        />
+      <Suspense fallback={null}>
+        <LazyBubbleChart {...props} />
+      </Suspense>
     );
 };
