+++ src/ts/components/charts/fragments/BarChart.tsx
@@ -1,59 +1,8 @@
 import { BarChart as MantineBarChart } from "@mantine/charts";
-import {
-    BarChartSeries,
-    BarChartType,
-} from "@mantine/charts/lib/BarChart/BarChart";
-import { MantineColor } from "@mantine/core";
-import { BoxProps } from "props/box";
-import { GridChartBaseProps } from "props/charts";
-import { DashBaseProps } from "props/dash";
-import { StylesApiProps } from "props/styles";
 import React, { useState, useRef } from "react";
 import { getClickData, isEventValid } from "../../../utils/charts";
 import { getLoadingState } from "../../../utils/dash3";
-
-interface Props
-    extends BoxProps,
-        GridChartBaseProps,
-        StylesApiProps,
-        DashBaseProps {
-    /** Data used to display chart */
-    data: Record<string, any>[];
-    /** An array of objects with `name` and `color` keys. Determines which data should be consumed from the `data` array. */
-    series: BarChartSeries[];
-    /** Controls how bars are positioned relative to each other, `'default'` by default */
-    type?: BarChartType;
-    /** Controls fill opacity of all bars, `1` by default */
-    fillOpacity?: number;
-    /** Fill of hovered bar section, by default value is based on color scheme */
-    cursorFill?: MantineColor;
-    /** Props passed down to recharts `BarChart` component */
-    barChartProps?: object;
-    /** Props passed down to recharts `Bar` component */
-    barProps?: object;
-    /** Additional components that are rendered inside recharts `BarChart` component */
-    children?: React.ReactNode;
-    /** Click data */
-    clickData?: Record<string, any>;
-    /** Hover data */
-    hoverData?: Record<string, any>;
-    /** Name of the series that was clicked */
-    clickSeriesName?: Record<string, any>;
-    /** Name of the series that is hovered*/
-    hoverSeriesName?: Record<string, any>;
-    /** Determines whether a label with bar value should be displayed on top of each bar,
-     incompatible with type="stacked" and type="percent", false by default */
-    withBarValueLabel?: boolean;
-    /**Determines whether a hovered series is highlighted. False by default. Mirrors the behaviour when hovering about chart legend items*/
-    highlightHover?: boolean
-    /** Sets minimum height of the bar in px, `0` by default */
-    minBarSize?: number;
-    /** Maximum bar width in px */
-    maxBarWidth?: number;
-    /** Controls color of the bar label, by default the value is determined by the chart orientation */
-    barLabelColor?: MantineColor;
-}
-
+import { Props }  from "../BarChart"
 
 /** BarChart */
 const BarChart = ({
